backend/requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
certifi==2025.4.26; python_version >= '3.6'
charset-normalizer==3.4.2; python_version >= '3.7'
click==8.2.1; python_version >= '3.10'
fastapi==0.115.12; python_version >= '3.8'
h11==0.16.0; python_version >= '3.8'
httptools==0.6.4
idna==3.10; python_version >= '3.6'
pydantic==2.11.5; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
python-dotenv==1.1.0; python_version >= '3.9'
pyyaml==6.0.2; python_version >= '3.8'
openai>=1.0.0; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.46.2; python_version >= '3.9'
typing-extensions==4.13.2; python_version >= '3.8'
typing-inspection==0.4.1; python_version >= '3.9'
urllib3[socks]==2.4.0; python_version >= '3.9'
uvicorn[standard]==0.34.2; python_version >= '3.9'
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1; python_version >= '3.9'
